Break seamly2d/me into separate Casks w/dependency.
Turns out that the nested container stanza just does not have any
provisions for specifying multiple containers nested in the main
container. That's annoying, but I also don't feel like navigating the
processes of coming up with a patch for the code and then doing a PR.
So, instead, screw it, we'll do it this way.